fix(core/yjk): fix type error
