+++ packages/outreader-core/src/interfaces/index.ts
@@ -3,4 +3,4 @@ export * from './wdisp.interface';
 export * from './wmass.interface';
 export * from './wv02q.interface';
 export * from './wzq.interface';
-export * from './concreteSteel.interface';
+export * from './concrete-steel.interface';
