+++ packages/outreader-core/src/interfaces/structure.interface.ts
@@ -2,7 +2,7 @@ import { IWdisp } from './wdisp.interface';
 import { IWmass } from './wmass.interface';
 import { IWv02q } from './wv02q.interface';
 import { IWzq } from './wzq.interface';
-import { IConcreteSteel } from './concreteSteel.interface';
+import { IConcreteSteel } from './concrete-steel.interface';
 
 export interface IStructure {
   hash: string;
