--- packages/outreader-yjk/src/concreteSteel/index.ts
@@ -1,1 +1,0 @@
-export * from './concreteSteel';
