+++ packages/outreader-yjk/src/index.ts
@@ -3,4 +3,4 @@ export * from './wdisp';
 export * from './wmass';
 export * from './wv02q';
 export * from './wzq';
-export * from './concreteSteel';
+export * from './concrete-steel';
